+++ src/scalib/config/__init__.py
@@ -143,4 +143,13 @@ def default_config(**kwargs):
 
 def get_config():
     """Get the current config, to give to Rust functions."""
-    return _current.get()
+    try:
+        return _current.get()
+    except LookupError:
+        # TODO: use add_note on the orignal exception once we drop support for python 3.10.
+        raise Exception(
+            "Using SCALib from a thread without contextvars propagation is not supported.\n"
+            + "We provide contextvars-aware replacements for ThreadPoolExecutor and Thread\n"
+            + "at scalib.tools.ContextExecutor and scalib.tools.ContextThread.\n"
+            + "See https://github.com/simple-crypto/SCALib/issues/89 for more information."
+        )